move sentinel
